+++ packages/cli/src/cli/commands/actors/index.ts
@@ -1,10 +1,8 @@
 import { Command } from "commander";
 import { getDeployCommand } from "./deploy.js";
-import { getNewCommand } from "./new.js";
 
-export function getActorCommand(): Command {
-  return new Command("actor")
+export function getActorsCommand(): Command {
+  return new Command("actors")
     .description("Manage actors")
-    .addCommand(getNewCommand())
     .addCommand(getDeployCommand());
 }